--- routes/users.js
@@ -1,18 +1,0 @@
-"use strict";
-
-const express = require('express');
-const router  = express.Router();
-
-module.exports = (knex) => {
-
-  router.get("/", (req, res) => {
-    knex
-      .select("*")
-      .from("users")
-      .then((results) => {
-        res.json(results);
-    });
-  });
-
-  return router;
-}
