+++ views/index.ejs
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <title>Home Page</title>
+    <title>Schoodle - Home Page</title>
 
     <link rel="stylesheet" href="/vendor/normalize-4.1.1.css" type="text/css" />
     <link rel="stylesheet" href="/vendor/border-box.css" type="text/css" />
@@ -10,8 +10,42 @@
     <script type="text/javascript" src="/vendor/jquery-3.0.0.js"></script>
     <script type="text/javascript" src="/scripts/app.js"></script>
   </head>
+  <header><% include partials/_header %></header>
 
   <body>
-    <h1>Hello World!</h1>
+      <h1>Welcome to Schoodle</h1>
+
+      <form action="/api/events/create" method="POST">
+
+        <div class="eventInput">
+          <label type="text" >Event name:</label>
+          <input type="text" name="theEventName">
+        </div>
+
+
+      <label type="text">Event Description:</label>
+      <textarea name="theEventDescription" placeholder="Tell us about thie event." ></textarea>
+
+
+      <label type="text">Enter your name:</label>
+      <input type="text" name="theHostName">
+
+      <label type="text">Enter your email:</label>
+      <input type="email" name="theHostMail">
+
+      <div class="eventSetup">
+          <input type="datetime-local" id="getDate" name="eventTime">
+          <button id="setDate" type="button" name="addTime" >Add Timeslot</button>
+          <h1>Select time slots</h1>
+            <p class="timeZone">
+              <li>
+
+              </li>
+            </p>
+
+      </div>
+
+      <button id="createEvent" type="submit">Create the Event</button>
+    </form>
   </body>
 </html>
