+++ views/partials/_header.ejs
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head></head>
+  <body>
+      <form >
+        <h1 class=”websiteName”>Schoodle</h1>
+        <p>- Your mate for events</p>
+      </form>
+  </body>
+</html>
